+++ src/webserver/database/sql_to_json.rs
@@ -156,16 +156,6 @@ pub fn sql_nonnull_to_json<'r>(mut get_ref: impl FnMut() -> sqlx::any::AnyValueR
     }
 }
 
-/// Takes the first column of a row and converts it to a string.
-pub fn row_to_string(row: &AnyRow) -> Option<String> {
-    let col = row.columns().first()?;
-    match sql_to_json(row, col) {
-        serde_json::Value::String(s) => Some(s),
-        serde_json::Value::Null => None,
-        other => Some(other.to_string()),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::app_config::tests::test_database_url;