+++ tests/sql_test_files/component_rendering/error_set_scalar_query_returned_more_than_one_column.sql
@@ -1,0 +1,3 @@
+set x = (select 1 as a, 2 as b);
+
+select 'text' as component, 'It works !' as contents;