+++ tests/uploads/upload_file_runsql_test.sql
@@ -1,1 +1,2 @@
-select 'dynamic' as component, sqlpage.run_sql('tests/uploads/upload_file_test.sql') as properties;
+select 'dynamic' as component,
+    coalesce(sqlpage.run_sql('tests/uploads/upload_file_test.sql'), '[]') as properties;